Separate format and sort tests into two different files
